+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - Added `\Wizaplace\SDK\Catalog\ProductAttribute::getValueIds`
 - Added `\Wizaplace\SDK\Catalog\AttributeVariant::getDescription`
+- Added `\Wizaplace\SDK\Catalog\SearchProductAttribute::getType`
+- Added `\Wizaplace\SDK\Catalog\AttributeType::FREE`
 
 ### Bugfixes
 </details>